requirements.txt
opencv_python==4.5.4.60
protobuf==3.19.1
python_telegram_bot==13.8.1
scikit_image==0.19.0
google-cloud-vision
